Move mapping-transform to mapping-transform-script
Its more descriptive.
