+++ plugins/mapping-transform-script/licenses/no_deps.txt
