+++ plugins/mapping-transform-script/rest-api-spec/test/mapping_transform_script/10_basic.yaml
