+++ plugins/pom.xml
@@ -434,10 +434,10 @@
         <module>cloud-azure</module>
         <module>cloud-aws</module>
         <module>delete-by-query</module>
-        <module>mapping-transform</module>
         <module>discovery-multicast</module>
         <module>lang-python</module>
         <module>lang-javascript</module>
+        <module>mapping-transform-script</module>
         <module>mapper-murmur3</module>
         <module>mapper-size</module>
         <module>jvm-example</module>
